pyproject.toml
[tool.poetry]
name = "trading-api"
version = "0.1.0"
description = "manage posiotions from different terminals"
authors = ["Elizaveta"]
readme = "README.md"
packages = [{include = "positions"}]

[tool.poetry.dependencies]
python = "^3.10"
pymongo = "^4.7.2"
python-dotenv = "^1.0.1"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "positions.main:main"
